feat: migrate various base commands
