+++ instagram-ts/source/cli.ts
@@ -1,0 +1,8 @@
+#!/usr/bin/env node
+import Pastel from 'pastel';
+
+const app = new Pastel({
+	importMeta: import.meta,
+});
+
+await app.run();
