--- instagram-ts/source/cli.tsx
@@ -1,8 +1,0 @@
-#!/usr/bin/env node
-import {runCli} from './commands.js';
-
-// Run the CLI
-runCli().catch(error => {
-	console.error('Fatal error:', error);
-	process.exit(1);
-});
