+++ instagram-ts/source/client.ts
@@ -1,65 +1,165 @@
 import {IgApiClient} from 'instagram-private-api';
 import {SessionManager} from './session.js';
+import {ConfigManager} from './config.js';
 import type {Thread, Message, User} from './types/instagram.js';
+
+export interface LoginResult {
+	success: boolean;
+	error?: string;
+	username?: string;
+}
 
 export class InstagramClient {
 	private ig: IgApiClient;
 	private sessionManager: SessionManager | null = null;
+	private configManager: ConfigManager;
+	private username: string | null = null;
 
-	constructor() {
+	constructor(username?: string) {
 		this.ig = new IgApiClient();
+		this.configManager = ConfigManager.getInstance();
+
+		if (username) {
+			this.username = username;
+			this.sessionManager = new SessionManager(username);
+		}
 	}
 
-	async login(username: string, password: string): Promise<boolean> {
+	public async login(
+		username: string,
+		password: string,
+		verificationCode?: string,
+	): Promise<LoginResult> {
 		try {
+			this.username = username;
 			this.sessionManager = new SessionManager(username);
+
+			// Check if we should load existing session first
+			const sessionExists = await this.sessionManager.sessionExists();
+			if (sessionExists) {
+				try {
+					const sessionData = await this.sessionManager.loadSession();
+					if (sessionData) {
+						await this.ig.state.deserialize(sessionData);
+						// Preserve device UUIDs from old session
+						const oldSettings = await this.ig.state.serialize();
+						this.ig.state.generateDevice(username);
+						if (oldSettings.deviceString) {
+							// Restore device string to maintain consistency
+							await this.ig.state.deserialize({
+								...sessionData,
+								deviceString: oldSettings.deviceString,
+							});
+						}
+					}
+				} catch (error) {
+					console.log('Could not load existing session, creating new one');
+				}
+			}
 
 			// Generate device for this username
 			this.ig.state.generateDevice(username);
 
-			// Subscribe to save state after each request
+			// Set up request listener to save session after each request
 			this.ig.request.end$.subscribe(async () => {
-				const serialized = await this.ig.state.serialize();
-				delete serialized.constants;
-				await this.sessionManager!.saveSession(serialized);
+				await this.saveSessionState();
 			});
 
-			// Attempt login
-			await this.ig.account.login(username, password);
-			return true;
+			// Perform login
+			await this.ig.simulate.preLoginFlow();
+
+			if (verificationCode) {
+				// If 2FA is needed, this would need to be handled differently
+				// For now, we'll assume the verification code is provided upfront
+				await this.ig.account.login(username, password);
+			} else {
+				await this.ig.account.login(username, password);
+			}
+
+			// Save session and update config
+			await this.saveSessionState();
+			await this.configManager.set('login.currentUsername', username);
+
+			// Set as default username if none exists
+			const defaultUsername = this.configManager.get<string>(
+				'login.defaultUsername',
+			);
+			if (!defaultUsername) {
+				await this.configManager.set('login.defaultUsername', username);
+			}
+
+			return {success: true, username};
 		} catch (error) {
 			console.error('Login failed:', error);
-			return false;
+			return {
+				success: false,
+				error: error instanceof Error ? error.message : 'Unknown login error',
+			};
 		}
 	}
 
-	async restoreSession(username: string): Promise<boolean> {
-		try {
-			this.sessionManager = new SessionManager(username);
+	public async loginBySession(): Promise<LoginResult> {
+		if (!this.sessionManager) {
+			return {success: false, error: 'No session manager initialized'};
+		}
 
+		try {
 			const sessionData = await this.sessionManager.loadSession();
 			if (!sessionData) {
-				return false;
+				return {success: false, error: 'No session file found'};
 			}
 
-			this.ig.state.generateDevice(username);
-
-			// Subscribe to save state after each request
-			this.ig.request.end$.subscribe(async () => {
-				const serialized = await this.ig.state.serialize();
-				delete serialized.constants;
-				await this.sessionManager!.saveSession(serialized);
-			});
-
+			// Deserialize the session state
 			await this.ig.state.deserialize(sessionData);
 
-			// Test if the session is still valid
-			await this.ig.user.info(this.ig.state.cookieUserId);
-			return true;
+			// Test if session is valid by making a simple request
+			const currentUser = await this.ig.account.currentUser();
+			this.username = currentUser.username;
+
+			// Save the session after successful login
+			await this.saveSessionState();
+			await this.configManager.set('login.currentUsername', this.username);
+
+			return {success: true, username: this.username};
 		} catch (error) {
-			console.error('Session restoration failed:', error);
-			return false;
+			console.error('Failed to login with session:', error);
+			return {
+				success: false,
+				error: error instanceof Error ? error.message : 'Unknown session error',
+			};
 		}
+	}
+
+	private async saveSessionState(): Promise<void> {
+		if (!this.sessionManager) {
+			return;
+		}
+
+		try {
+			const serialized = await this.ig.state.serialize();
+			await this.sessionManager.saveSession(serialized);
+		} catch (error) {
+			console.error('Error saving session state:', error);
+		}
+	}
+
+	public async logout(): Promise<void> {
+		try {
+			if (this.sessionManager) {
+				await this.sessionManager.deleteSession();
+			}
+			await this.configManager.set('login.currentUsername', null);
+		} catch (error) {
+			console.error('Error during logout:', error);
+		}
+	}
+
+	public getInstagramClient(): IgApiClient {
+		return this.ig;
+	}
+
+	public getUsername(): string | null {
+		return this.username;
 	}
 
 	async getThreads(): Promise<Thread[]> {
@@ -80,26 +180,32 @@ export class InstagramClient {
 		}
 	}
 
-	async getMessages(threadId: string): Promise<Message[]> {
+	async getMessages(
+		threadId: string,
+		cursor?: string,
+	): Promise<{messages: Message[]; cursor: string | undefined}> {
 		try {
 			const thread = this.ig.feed.directThread({
 				thread_id: threadId,
-				oldest_cursor: '',
+				oldest_cursor: cursor || '',
 			});
 			const items = await thread.items();
 
-			return items
-				.filter(item => item.item_type === 'text')
-				.map(item => ({
-					id: item.item_id,
-					text: item.text || '',
-					timestamp: new Date((item.timestamp as any) / 1000),
-					userId: item.user_id.toString(),
-					username: this.getUsernameFromId(item.user_id),
-					isOutgoing: item.user_id.toString() === this.ig.state.cookieUserId,
-					threadId: threadId,
-				}))
-				.reverse(); // Show newest messages at bottom
+			return {
+				messages: items
+					.filter(item => item.item_type === 'text')
+					.map(item => ({
+						id: item.item_id,
+						text: item.text || '',
+						timestamp: new Date((item.timestamp as any) / 1000),
+						userId: item.user_id.toString(),
+						username: this.getUsernameFromId(item.user_id),
+						isOutgoing: item.user_id.toString() === this.ig.state.cookieUserId,
+						threadId: threadId,
+					}))
+					.reverse(), // Show newest messages at bottom
+				cursor: (thread as any).oldestCursor,
+			};
 		} catch (error) {
 			console.error('Failed to fetch messages:', error);
 			throw error;
